+++ cooking_book/src/main/kotlin/com/example/cooking_book/repository/StepRepository.kt
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 
-interface StepRepository : JpaRepository<Step, Int> {
+interface StepRepository : JpaRepository<Step, Long> {
     @Modifying
     @Transactional
     @Query("UPDATE Step r SET r.stepDescription = :stepDescription, r.stepOrder = :stepOrder " +
